+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/snapshots/snap_test_assets.py
@@ -40,3 +40,126 @@ snapshots['TestAssetAwareEventLog.test_get_asset_key_not_found[asset_aware_insta
         '__typename': 'AssetNotFoundError'
     }
 }
+
+snapshots['TestAssetAwareEventLog.test_get_all_asset_keys[asset_aware_instance_in_process_env] 1'] = {
+    'assetsOrError': {
+        '__typename': 'AssetConnection',
+        'nodes': [
+            {
+                'key': {
+                    'path': [
+                        'a'
+                    ]
+                }
+            },
+            {
+                'key': {
+                    'path': [
+                        'b'
+                    ]
+                }
+            },
+            {
+                'key': {
+                    'path': [
+                        'c'
+                    ]
+                }
+            }
+        ]
+    }
+}
+
+snapshots['TestAssetAwareEventLog.test_get_prefixed_asset_keys[in_memory_instance_in_process_env] 1'] = {
+    'assetsOrError': {
+        '__typename': 'AssetConnection',
+        'nodes': [
+            {
+                'key': {
+                    'path': [
+                        'a'
+                    ]
+                }
+            }
+        ]
+    }
+}
+
+snapshots['TestAssetAwareEventLog.test_get_prefixed_asset_keys[asset_aware_instance_in_process_env] 1'] = {
+    'assetsOrError': {
+        '__typename': 'AssetConnection',
+        'nodes': [
+            {
+                'key': {
+                    'path': [
+                        'a'
+                    ]
+                }
+            }
+        ]
+    }
+}
+
+snapshots['TestAssetAwareEventLog.test_get_asset_key_materialization[in_memory_instance_in_process_env] 1'] = {
+    'assetOrError': {
+        'assetMaterializations': [
+            {
+                'materializationEvent': {
+                    'materialization': {
+                        'label': 'a'
+                    }
+                }
+            }
+        ]
+    }
+}
+
+snapshots['TestAssetAwareEventLog.test_get_asset_key_materialization[asset_aware_instance_in_process_env] 1'] = {
+    'assetOrError': {
+        'assetMaterializations': [
+            {
+                'materializationEvent': {
+                    'materialization': {
+                        'label': 'a'
+                    }
+                }
+            }
+        ]
+    }
+}
+
+snapshots['TestAssetAwareEventLog.test_get_asset_key_not_found[in_memory_instance_in_process_env] 1'] = {
+    'assetOrError': {
+        '__typename': 'AssetNotFoundError'
+    }
+}
+
+snapshots['TestAssetAwareEventLog.test_get_partitioned_asset_key_materialization[in_memory_instance_in_process_env] 1'] = {
+    'assetOrError': {
+        'assetMaterializations': [
+            {
+                'materializationEvent': {
+                    'materialization': {
+                        'label': 'a'
+                    }
+                },
+                'partition': 'partition_1'
+            }
+        ]
+    }
+}
+
+snapshots['TestAssetAwareEventLog.test_get_partitioned_asset_key_materialization[asset_aware_instance_in_process_env] 1'] = {
+    'assetOrError': {
+        'assetMaterializations': [
+            {
+                'materializationEvent': {
+                    'materialization': {
+                        'label': 'a'
+                    }
+                },
+                'partition': 'partition_1'
+            }
+        ]
+    }
+}