+++ python_modules/dagster/dagster_tests/core_tests/config_types_tests/test_config_default_required.py
@@ -1,4 +1,4 @@
-from dagster import Field, execute_solid, solid
+from dagster import Field, Noneable, execute_solid, solid
 
 
 def test_default_implies_not_required_field_correct():
@@ -17,6 +17,29 @@ def test_default_implies_not_required_execute_solid():
     def return_default_to_one(context):
         return context.solid_config["default_to_one"]
 
-    # This currently throws. It should not require configuration as the
-    # the default value should make it *not* required.
     execute_solid(return_default_to_one)
+
+
+def test_scalar_field_defaults():
+    assert Field(int).is_required is True
+    assert Field(Noneable(int)).is_required is False
+    assert Field(Noneable(int)).default_value is None
+
+
+def test_noneable_shaped_field_defaults():
+    schema = {"an_int": int}
+    assert Field(schema).is_required is True
+    assert Field(Noneable(schema)).is_required is False
+    assert Field(Noneable(schema)).default_value is None
+
+
+def test_noneable_string_in_solid():
+    executed = {}
+
+    @solid(config_schema=Noneable(int))
+    def default_noneable_int(context):
+        assert context.solid_config is None
+        executed["yes"] = True
+
+    execute_solid(default_noneable_int)
+    assert executed["yes"]